+++ module_groups/ecs_fargate_with_elb/variables.tf
@@ -54,6 +54,21 @@ variable "health_check_timeout" {
   default     = 5
 }
 
+variable "health_check_interval" {
+  description = "Approximate amount of time, in seconds, between health checks of an individual target."
+  default     = 30
+}
+
+variable "health_check_healthy_threshold" {
+  description = "Number of consecutive health check successes required before considering a target healthy."
+  default     = 3
+}
+
+variable "health_check_unhealthy_threshold" {
+  description = "Number of consecutive health check failures required before considering a target unhealthy. "
+  default     = 3
+}
+
 variable "internal" {
   description = "An Internet-facing load balancer routes requests from clients over the Internet to targets. An internal load balancer routes requests from clients to targets using private IP addresses."
   default     = false