+++ modules/public_load_balancer/main.tf
@@ -35,9 +35,12 @@ resource "aws_lb_target_group" "target_group" {
   slow_start           = var.slow_start
 
   health_check {
-    path    = var.health_check_path
-    timeout = var.health_check_timeout
-    matcher = "200"
+    path                = var.health_check_path
+    timeout             = var.health_check_timeout
+    interval            = var.health_check_interval
+    healthy_threshold   = var.health_check_healthy_threshold
+    unhealthy_threshold = var.health_check_unhealthy_threshold
+    matcher             = "200"
   }
 
   stickiness {